#1602-nifti-conversion - correct rabbitMQ logic in the application + download naming with seconds

+++ shanoir-ng-datasets/src/main/java/org/shanoir/ng/bids/service/BIDSServiceImpl.java
@@ -192,7 +192,7 @@ public class BIDSServiceImpl implements BIDSService {
 					value = @Queue(value = RabbitMQConfiguration.BIDS_EVENT_QUEUE, durable = "true"),
 					exchange = @Exchange(value = RabbitMQConfiguration.EVENTS_EXCHANGE, ignoreDeclarationExceptions = "true",
 					autoDelete = "false", durable = "true", type=ExchangeTypes.TOPIC))
-	}
+	}, containerFactory = "multipleConsumersFactory"
 			)
 	public void deleteBids(String eventAsString) {
 		ShanoirEvent event;
@@ -203,6 +203,9 @@ public class BIDSServiceImpl implements BIDSService {
 				return;
 			}
 			Study studyDeleted = studyRepo.findById(event.getStudyId()).orElse(null);
+			if (studyDeleted == null) {
+				return;
+			}
 			this.deleteBidsFolder(studyDeleted.getId(), studyDeleted.getName());
 		} catch (Exception e) {
 			LOG.error("ERROR when deleting BIDS folder: please delete it manually: {}", eventAsString, e);
